certificate_validation/requirements.txt
pymupdf
easyocr
pillow
pyzbar
requests
transformers
torch
torchvision
torchaudio
rapidfuzz
pymongo
opencv-python-headless